refactor: Moved SpaceInvaders files to SpaceInvaders project
