+++ SpaceInvaders/Enums.cs
@@ -1,0 +1,9 @@
+﻿namespace SpaceInvaders;
+
+public enum Keys
+{
+    Enter,
+    Left,
+    Right,
+    Up,
+}
